SO 입력 값 변경 / UI 적용중

+++ Assets/Scripts/ScriptableObj/Forge/Rebuild/Attacks/GuidedMissile/GuidedMissileAtkDamageSO.cs
@@ -2,13 +2,13 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "GuidedMissileAtkDamageSO", menuName = "ScriptableObjects/Forge/Attacks/GuidedMissile/GuidedMissileAtkDamageSO", order = 1)]
 public class GuidedMissileAtkDamageSO : BaseForgeSO
 {
-    public int AtkDamage;
+    public float AtkDamage;
     
     protected override ForgeId GetForgeId() => ForgeId.GuidedMissileAtkDamage;
     
     public override void Apply()
     {
         if (Managers.Instance?.turretActivationManager == null) return;
-        Managers.Instance.turretActivationManager.AddMissileDamage((float)AtkDamage);
+        Managers.Instance.turretActivationManager.AddMissileDamage(AtkDamage);
     }
 }
