+++ Assets/Scripts/ScriptableObj/Forge/Rebuild/Attacks/GuidedMissile/GuidedMissileAtkSpeedSO.cs
@@ -2,13 +2,13 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "GuidedMissileAtkSpeedSO", menuName = "ScriptableObjects/Forge/Attacks/GuidedMissile/GuidedMissileAtkSpeedSO", order = 1)]
 public class GuidedMissileAtkSpeedSO : BaseForgeSO
 {
-    public int AtkSpeed;
+    public float AtkSpeed;
     
     protected override ForgeId GetForgeId() => ForgeId.GuidedMissileAtkSpeed;
     
     public override void Apply()
     {
         if (Managers.Instance?.turretActivationManager == null) return;
-        Managers.Instance.turretActivationManager.AddMissileInterval((float)AtkSpeed);
+        Managers.Instance.turretActivationManager.AddMissileInterval(AtkSpeed);
     }
 }
