+++ Assets/Scripts/ScriptableObj/Forge/Rebuild/Planet/PlanetCoreMaxHpSO.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PlanetCoreMaxHpSO", menuName = "ScriptableObjects/Forge/Planet/PlanetCoreMaxHpSO", order = 1)]
 public class PlanetCoreMaxHpSO : BaseForgeSO
 {
-    public int CoreMaxHp;
+    public float CoreMaxHp;
     
     protected override ForgeId GetForgeId() => ForgeId.PlanetCoreMaxHp;
