+++ Assets/Scripts/ScriptableObj/Forge/Rebuild/SpaceShip/Move/SpaceShipMoveOrePerSlowSO.cs
@@ -2,13 +2,13 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "SpaceShipMoveOrePerSlowSO", menuName = "ScriptableObjects/Forge/SpaceShip/Move/SpaceShipMoveOrePerSlowSO", order = 1)]
 public class SpaceShipMoveOrePerSlowSO : BaseForgeSO
 {
-    public int OrePerSlow;
+    public float OrePerSlow;
     
     protected override ForgeId GetForgeId() => ForgeId.SpaceShipMoveOrePerSlow;
     
     public override void Apply()
     {
         if (Managers.Instance?.spaceshipMotor == null) return;
-        Managers.Instance.spaceshipMotor.AddThrustReductionPerOre((float)OrePerSlow);
+        Managers.Instance.spaceshipMotor.AddThrustReductionPerOre(OrePerSlow);
     }
 }
